Refactor Guessing, Guess components, extract functions.
